Move gitforge into submodule
